+++ client/src/api/login/login.ts
@@ -1,6 +1,5 @@
 // TODO: 로그인 하고 아이디와 비밀번호 일치 여부 체크 /user/login
 
-import axios from 'axios';
 import { API } from '../api.const';
 
 interface LoginResponse {
@@ -16,6 +15,7 @@ const loginUser = async (
 ): Promise<{ success: boolean; nickname?: string; error?: string }> => {
   try {
     // 서버에 로그인 요청을 보내고, 사용자 정보를 받아옴
+    // TODO: API = 'http://123123123' 포함
     const response = await fetch(`http://${API}/user/login`, {
       method: 'POST',
       headers: {
