+++ client/src/components/myPage/MyPostList.tsx
@@ -1,6 +1,5 @@
-
 import { Post, PostType } from '../post';
-import {PostList} from './styles'
+import { PostList } from './MyPage.style';
 
 export type PostsType = PostType[];
