+++ client/src/components/new/NewPostList.tsx
@@ -1,14 +1,13 @@
-import { PostType } from '../post';
-import { Post } from '../post';
-import { PostList } from './styles';
 import { useEffect, useState } from 'react';
-import { fetchNewPosts, fetchNextPage } from '../../api/post/newPost';
+import { fetchNewPosts, fetchNextPage } from '../../api/post/newPost'; // TODO: '../../api'로 불러올 수 있게 수정
+import { Post, PostType } from '../post';
+import { PostList } from './styles';
 
 interface NewPostListProps {
   posts: PostType[]; // props 형식 수정
 }
 
-export const NewPostList: React.FC<NewPostListProps> = ({ posts: initialPosts }) => {
+export const NewPostList = ({ posts: initialPosts }: NewPostListProps) => {
   const [posts, setPosts] = useState<PostType[]>(initialPosts);
   const [page, setPage] = useState(1);
 
@@ -28,7 +27,7 @@ export const NewPostList: React.FC<NewPostListProps> = ({ posts: initialPosts })
   useEffect(() => {
     // 최초 렌더링 시 초기 게시물을 가져옴
     fetchPosts(page);
-  }, []);
+  }, [page]);
 
   const fetchPosts = (pageNum: number) => {
     fetchNextPage(pageNum).then(nextPosts => {
@@ -52,12 +51,12 @@ export const NewPostList: React.FC<NewPostListProps> = ({ posts: initialPosts })
 
     window.addEventListener('scroll', handleScroll);
     return () => window.removeEventListener('scroll', handleScroll);
-  }, [page]); // 페이지 변경 시마다 이벤트 리스너를 업데이트
+  }, []); // 페이지 변경 시마다 이벤트 리스너를 업데이트
 
   return (
     <PostList>
       {posts.length === 0 ? (
-        <div>게시글이 없습니다.</div>
+        <div style={{ padding: '50px 0', marginBottom: '30' }}>게시글이 없습니다.</div>
       ) : (
         posts.map(post => <Post key={post.id} post={post} />)
       )}
